+++ CHANGELOG.md
@@ -5,6 +5,10 @@ This a global changelog that outlines overal changes in LiteMol.
 Each component also has its separate changelog 
 ([Core](src/lib/Core/CHANGELOG.md), [Visualization](src/lib/Visualization/CHANGELOG.md), [Bootstrap](src/lib/Bootstrap/CHANGELOG.md), [Plugin](src/lib/Plugin/CHANGELOG.md), [Viewer](src/Viewer/CHANGELOG.md)).
 
+2.4.1
+-----------
+
+* Fixed screenshot support (Chrome changed behaviour of data URIs).
 
 2.4.0-preview
 -----------